+++ R/utils.R
@@ -307,17 +307,3 @@ quo_invert <- function(call) {
 quo_is_constant <- function(quo) {
   is_reference(quo_get_env(quo), empty_env())
 }
-
-vec_simplify <- function(x) {
-  if (!vctrs::vec_is_list(x)) {
-    return(x)
-  }
-  if (!every(x, ~ vctrs::vec_is(.x) && vctrs::vec_size(.x) == 1L)) {
-    return(x)
-  }
-
-  tryCatch(
-    vctrs_error_incompatible_type = function(...) x,
-    vctrs::vec_c(!!!x)
-  )
-}